# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  message: string;
}

export type Role = typeof Role[keyof typeof Role];


export const Role = {
  user: 'user',
  admin: 'admin',
} as const;

export type ConversationStatus = typeof ConversationStatus[keyof typeof ConversationStatus];


export const ConversationStatus = {
  awaiting_reply: 'awaiting_reply',
  responded: 'responded',
  urgent: 'urgent',
  archived: 'archived',
} as const;

export type SenderType = typeof SenderType[keyof typeof SenderType];


export const SenderType = {
  user: 'user',
  admin: 'admin',
} as const;

export type NotificationType = typeof NotificationType[keyof typeof NotificationType];


export const NotificationType = {
  reply: 'reply',
  resource: 'resource',
  maintenance: 'maintenance',
} as const;

export type ResourceType = typeof ResourceType[keyof typeof ResourceType];


export const ResourceType = {
  article: 'article',
  book: 'book',
  video: 'video',
  podcast: 'podcast',
  verse: 'verse',
  exercise: 'exercise',
  prompt: 'prompt',
  tip: 'tip',
} as const;

export type ResourceTopic = typeof ResourceTopic[keyof typeof ResourceTopic];


export const ResourceTopic = {
  anxiety: 'anxiety',
  grief: 'grief',
  purpose: 'purpose',
  relationships: 'relationships',
  depression: 'depression',
  stress: 'stress',
  faith: 'faith',
  healing: 'healing',
  'self-worth': 'self-worth',
} as const;

export interface User {
  id: string;
  name: string;
  email: string;
  role: Role;
  emailVerified: boolean;
  createdAt: string;
}

export interface RegisterProfileRequest {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
}

export interface Message {
  id: string;
  conversationId: string;
  senderType: SenderType;
  senderName: string;
  body: string;
  riskCategories?: string[] | null;
  createdAt: string;
}

export interface CreateMessageRequest {
  /**
     * @minLength 1
     * @maxLength 20000
     */
  body: string;
}

export interface CreateConversationRequest {
  /**
     * @minLength 1
     * @maxLength 20000
     */
  body: string;
}

export interface Conversation {
  id: string;
  status: ConversationStatus;
  tags: string[];
  preview: string;
  lastMessageAt: string;
  createdAt: string;
}

export type ConversationDetail = Conversation & {
  messages: Message[];
};

export interface ConversationUser {
  id: string;
  name: string;
  email: string;
}

export type AdminConversation = Conversation & {
  user: ConversationUser;
};

export type AdminConversationDetail = AdminConversation & {
  messages: Message[];
};

export interface PatchConversationRequest {
  status?: ConversationStatus;
  tags?: string[];
}

export interface Notification {
  id: string;
  type: NotificationType;
  title: string;
  body: string;
  conversationId: string | null;
  read: boolean;
  createdAt: string;
}

export interface Resource {
  id: string;
  title: string;
  description: string;
  type: ResourceType;
  topic: ResourceTopic;
  url: string | null;
  body: string | null;
  createdAt: string;
}

export interface CreateResourceRequest {
  /**
     * @minLength 1
     * @maxLength 200
     */
  title: string;
  /**
     * @minLength 1
     * @maxLength 500
     */
  description: string;
  type: ResourceType;
  topic: ResourceTopic;
  url?: string | null;
  body?: string | null;
}

export interface UpdateResourceRequest {
  /**
     * @minLength 1
     * @maxLength 200
     */
  title?: string;
  /**
     * @minLength 1
     * @maxLength 500
     */
  description?: string;
  type?: ResourceType;
  topic?: ResourceTopic;
  url?: string | null;
  body?: string | null;
}

export interface Analytics {
  totalUsers: number;
  activeUsers30d: number;
  newRegistrations7d: number;
  totalConversations: number;
  awaitingReplyCount: number;
  urgentCount: number;
  respondedCount: number;
  archivedCount: number;
  avgResponseTimeMinutes: number | null;
  resourceViews30d: number;
}

/**
 * Invalid request
 */
export type BadRequestResponse = Error;

/**
 * Not signed in
 */
export type UnauthorizedResponse = Error;

/**
 * Signed in but not allowed
 */
export type ForbiddenResponse = Error;

/**
 * Not found
 */
export type NotFoundResponse = Error;

export type SearchParameter = string;

export type ListConversationsParams = {
search?: SearchParameter;
};

export type ListResourcesParams = {
search?: SearchParameter;
topic?: ResourceTopic;
type?: ResourceType;
};

export type AdminListConversationsParams = {
search?: SearchParameter;
status?: ConversationStatus;
tag?: string;
};

export type AdminListResourcesParams = {
search?: SearchParameter;
topic?: ResourceTopic;
type?: ResourceType;
};