# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/profile:
    post:
      operationId: registerProfile
      tags: [auth]
      summary: Create or update the signed-in Supabase Auth user's Tikvah profile
      description: >
        Called once, right after Supabase `signUp`/`signInWithPassword` returns a
        session, so the profile (name, role) exists before the app relies on it.
        Idempotent — safe to call more than once; a later call updates `name`.
        Requests without a valid Supabase access token are rejected, so this can
        only run once a session exists (it may not exist immediately after signup
        if email confirmation is required — the profile is then created lazily,
        the first time an authenticated request succeeds).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterProfileRequest"
      responses:
        "200":
          description: Profile created or updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Not signed in
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    Role:
      type: string
      enum: [user, admin]
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/Role"
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, emailVerified, createdAt]
    RegisterProfileRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
      required: [name]
...