# lib/api-zod/src/generated/types/registerProfileRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterProfileRequest {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
}